Add all the results
Added and formatted the results for the information needs and the
queries. Added two scripts to extract and format the data from the
data files. Switched all instances of okapi and cosine similarity to
LaTeX macros, so we can change formatting easily. Did one count for
info need correctness. Started work on formatting info needs with their
queries; I'm still not sure of the best way to present that data.